Estable a hub stable and instance id
In order to be able to cleanup old services for a hub, we either need to
destroy those records eagerly when the activity stream ends or destroy
them on a hub reconnect. The issue with deing it eagerly is that all
manner of things can happen to cut the activity stream and if we nuke
the record set immediately, we'll get a lot of unnecessary service
record bouncing.

So instead we introduce a separate id along with the hub, a stable id.

All normal uses use the current id (now called the instance id) to
associate records with hubs as well as with network location records.
But the hub presents it's stable id at bootstarp time which allows the
system to see if there is a set of services from that stable ids
previous instance id. And if so, delete those records.

This will require a separate process to cleanup records from hubs that
are truly gone, that comes next.