Prevent data mutation bug
The old version was manipulating the LabelSet that is stored between
requests if a deployment URL is used, which results in the next request
not working because the deployment label will be duplicated in the Set.

The situation resolves itself when the LabelLinks are downloaded again
but is basically broken for a minute at a time. Let's not break it.