pkg/control: Pick up new hub image tag dynamically
This allows the HubImageTag value to be a file that is polled for
changes. This allows the kubernetes deployments to point the HubImageTag
value to a dynamically updated file from the configmap.

When we revise the hub/control connection model, we'll allow the hub to
push down a signal that the hubs need to pickup new config as well.