pkg/control: Attempt to fix table scan missing a row
It appears that the current code causes Postgresql to miss the recently
written row when scanning the label links table directly after adding
the new row, which causes the label link blob to miss that row entirely.

We think this is some MVCC behavior around making the row visible and so
we're moving this code into running inside the same transaction to force
postgresql to see the new row.