feat: disable insecureSkipVerify on by default
Addresses https://github.com/hashicorp/horizon/issues/49.
This commit adds the possibility to define, via a flag, whether
to enable or disable TLS certificate checks.

The default configuration (when `insecure` is false), is
actually using TLS but ignoring TLS validity. With this change,
the user will be able to specify via the `--insecure-skip-verify`
flag if he wants to skip the check, but we're switching the default
to be secure - exactly as one would expect when `--insecure` is
not selected.